refactor: moved files, change selectionSlice
