+++ src/features/select-item/model/selection-slice.ts
@@ -1,32 +1,41 @@
-import { createSlice } from '@reduxjs/toolkit';
+import type { Artwork } from 'entities/artwork';
 
-export const selectionSlice = createSlice({
+import { createSlice, type PayloadAction } from '@reduxjs/toolkit';
+
+const selectionSlice = createSlice({
   initialState: {
-    selected: [] as number[],
+    selected: [] as Artwork[],
   },
   name: 'selection',
-  reducers: (create) => ({
-    clearSelection: create.reducer((state) => {
+  reducers: {
+    clearSelection(state) {
       state.selected = [];
-    }),
-    toggleSelection: create.reducer<{ selectedId: number }>((state, action) => {
-      const { selectedId } = action.payload;
-      if (state.selected.includes(selectedId)) {
-        state.selected = state.selected.filter((id) => id !== selectedId);
+    },
+    toggleSelection(state, action: PayloadAction<{ selectedItem: Artwork }>) {
+      const { selectedItem } = action.payload;
+      const exists = state.selected.some((item) => item.id === selectedItem.id);
+      if (exists) {
+        state.selected = state.selected.filter(
+          (item) => item.id !== selectedItem.id
+        );
       } else {
-        state.selected.push(selectedId);
+        state.selected.push(selectedItem);
       }
-    }),
-  }),
+    },
+  },
   selectors: {
     getSelectedItems: (state) => state.selected,
-    isSelected: (state, id: number) => state.selected.includes(id),
+    isSelected: (state, id: number) =>
+      state.selected.some((item) => item.id === id),
   },
 });
 
-export const selectionReducer = selectionSlice.reducer;
 export const { clearSelection, toggleSelection } = selectionSlice.actions;
+
+export const selectionReducer = selectionSlice.reducer;
+
 export const { getSelectedItems, isSelected } = selectionSlice.selectors;
+
 export type SelectionInitialState = ReturnType<
   typeof selectionSlice.getInitialState
 >;
