+++ src/shared/lib/hooks.ts
@@ -1,6 +1,6 @@
 import { useDispatch, useSelector } from 'react-redux';
 
-import type { AppDispatch, RootState } from './store';
+import type { AppDispatch, RootState } from './store.ts';
 
 export const useAppDispatch = useDispatch.withTypes<AppDispatch>();
 export const useAppSelector = useSelector.withTypes<RootState>();
