+++ src/shared/lib/store.ts
